Revert "Lexer"
